# [ROX-11437] Batch resolve alerts operations to get performance boost