--- a/central/alert/datastore/bench_test.go
+++ b/central/alert/datastore/bench_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stackrox/rox/central/alert/datastore/internal/search"
 	rocksDBStore "github.com/stackrox/rox/central/alert/datastore/internal/store/rocksdb"
 	"github.com/stackrox/rox/central/globalindex"
+	"github.com/stackrox/rox/pkg/env"
 	"github.com/stackrox/rox/pkg/fixtures"
 	"github.com/stackrox/rox/pkg/rocksdb"
 	"github.com/stackrox/rox/pkg/sac"
@@ -17,8 +18,12 @@ import (
 )
 
 func BenchmarkDBs(b *testing.B) {
-	ctx := sac.WithAllAccess(context.Background())
+	if !env.PostgresDatastoreEnabled.BooleanSetting() {
+		b.Skipf("%q not set. Skip postgres test", env.PostgresDatastoreEnabled.EnvVar())
+		b.SkipNow()
+	}
 
+	ctx := sac.WithAllAccess(context.Background())
 	db, err := rocksdb.NewTemp("alert_bench_test")
 	require.NoError(b, err)
 	defer rocksdbtest.TearDownRocksDB(db)