--- a/central/alert/datastore/datastore_sac_test.go
+++ b/central/alert/datastore/datastore_sac_test.go
@@ -2,7 +2,6 @@ package datastore
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/blevesearch/bleve"
@@ -146,7 +145,6 @@ func (s *alertDatastoreSACTestSuite) TestMarkAlertStale() {
 		"(full) read-only cannot mark alert stale": {
 			scopeKey:      testutils.UnrestrictedReadCtx,
 			expectError:   true,
-			expectedFound: true,
 			expectedError: sac.ErrResourceAccessDenied,
 		},
 		"full read-write can mark alert stale": {
@@ -156,22 +154,22 @@ func (s *alertDatastoreSACTestSuite) TestMarkAlertStale() {
 		"full read-write on wrong cluster cannot mark alert stale": {
 			scopeKey:      testutils.Cluster1ReadWriteCtx,
 			expectError:   true,
-			expectedFound: false,
+			expectedError: sac.ErrResourceAccessDenied,
 		},
 		"read-write on wrong cluster and wrong namespace name cannot mark alert stale": {
 			scopeKey:      testutils.Cluster1NamespaceAReadWriteCtx,
 			expectError:   true,
-			expectedFound: false,
+			expectedError: sac.ErrResourceAccessDenied,
 		},
 		"read-write on wrong cluster and matching namespace name cannot mark alert stale": {
 			scopeKey:      testutils.Cluster1NamespaceBReadWriteCtx,
 			expectError:   true,
-			expectedFound: false,
+			expectedError: sac.ErrResourceAccessDenied,
 		},
 		"read-write on right cluster but wrong namespaces cannot mark alert stale": {
 			scopeKey:      testutils.Cluster2NamespacesACReadWriteCtx,
 			expectError:   true,
-			expectedFound: false,
+			expectedError: sac.ErrResourceAccessDenied,
 		},
 		"full read-write on right cluster can mark alert stale": {
 			scopeKey:    testutils.Cluster2ReadWriteCtx,
@@ -205,16 +203,12 @@ func (s *alertDatastoreSACTestSuite) TestMarkAlertStale() {
 			_, err = s.datastore.MarkAlertStaleBatch(ctx, alert1.GetId())
 			if !c.expectError {
 				s.NoError(err)
-			} else if !c.expectedFound {
-				s.Equal(fmt.Errorf("alert with id '%s' does not exist", alert1.GetId()), err)
 			} else {
 				s.Equal(c.expectedError, err)
 			}
 			_, err = s.datastore.MarkAlertStaleBatch(ctx, alert2.GetId())
 			if !c.expectError {
 				s.NoError(err)
-			} else if !c.expectedFound {
-				s.Equal(fmt.Errorf("alert with id '%s' does not exist", alert2.GetId()), err)
 			} else {
 				s.Equal(c.expectedError, err)
 			}