--- a/central/alert/datastore/internal/store/mocks/store.go
+++ b/central/alert/datastore/internal/store/mocks/store.go
@@ -158,6 +158,20 @@ func (mr *MockStoreMockRecorder) Upsert(ctx, alert interface{}) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Upsert", reflect.TypeOf((*MockStore)(nil).Upsert), ctx, alert)
 }
 
+// UpsertMany mocks base method.
+func (m *MockStore) UpsertMany(ctx context.Context, alerts []*storage.Alert) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "UpsertMany", ctx, alerts)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// UpsertMany indicates an expected call of UpsertMany.
+func (mr *MockStoreMockRecorder) UpsertMany(ctx, alerts interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpsertMany", reflect.TypeOf((*MockStore)(nil).UpsertMany), ctx, alerts)
+}
+
 // Walk mocks base method.
 func (m *MockStore) Walk(ctx context.Context, fn func(*storage.Alert) error) error {
 	m.ctrl.T.Helper()