--- a/central/pruning/pruning.go
+++ b/central/pruning/pruning.go
@@ -426,10 +426,8 @@ func (g *garbageCollectorImpl) markOrphanedAlertsAsResolved(deployments set.Froz
 		return
 	}
 	log.Infof("[Alert pruning] Found %d orphaned alerts", len(alertsToResolve))
-	for _, a := range alertsToResolve {
-		if err := g.alerts.MarkAlertStale(pruningCtx, a); err != nil {
-			log.Error(errors.Wrapf(err, "error marking alert %q as stale", a))
-		}
+	if _, err := g.alerts.MarkAlertStaleBatch(pruningCtx, alertsToResolve...); err != nil {
+		log.Error(errors.Wrap(err, "error marking alert as stale"))
 	}
 }