… central/alert/datastore/datastore_test.go
… central/cluster/datastore/datastore_impl_test.go
… central/detection/alertmanager/alert_manager_impl_test.go