## Description

- Added a method to mark bulk alerts resolved. 
- Unit tests compare resolving alerts in bulk vs resolving alerts individually.

Additionally:
- Do not perform read perm check when upserting records. If any read is being performed, it is only the side-effect of our business logic, but not expected by user
- Added write perm check to batch upsert. This will replace the service layer perm check.
- Fixed the unit test. The unit test was checking for wrong expectations. If the incoming context does not have correct permissions, an unauthorized access error should be returned, not a record not found error.


Benchmark results show significant improvement both on rocksdb (~100x) and postgres (~10x). 

```
# rocksdb:
BenchmarkDBs/rocksdb-8         	       1	2639426601 ns/op
BenchmarkDBs/markStale
BenchmarkDBs/markStale-8       	       1	231159944603 ns/op
BenchmarkDBs/markStaleBatch
BenchmarkDBs/markStaleBatch-8  	       1	2826356797 ns/op

# postgres:

BenchmarkDBsWithPostgres/markStale
BenchmarkDBsWithPostgres/markStale-8         	       1	13617595712 ns/op
BenchmarkDBsWithPostgres/markStaleBatch
BenchmarkDBsWithPostgres/markStaleBatch-8    	       1	1949332050 ns/op

PASS
```
## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed
Unit tests